Moved CreateParcel tests
